+++ src/Products/PloneMeeting/Meeting.py
@@ -19,7 +19,6 @@ from collective.contact.plonegroup.config import get_registry_organizations
 from collective.eeafaceted.dashboard.utils import enableFacetedDashboardFor
 from copy import deepcopy
 from DateTime import DateTime
-from DateTime.DateTime import _findLocalTimeZoneName
 from imio.helpers.cache import cleanRamCacheFor
 from imio.helpers.cache import invalidate_cachekey_volatile_for
 from imio.prettylink.interfaces import IPrettyLink
@@ -47,7 +46,6 @@ from Products.CMFCore.permissions import ReviewPortalContent
 from Products.CMFCore.utils import _checkPermission
 from Products.CMFDynamicViewFTI.browserdefault import BrowserDefaultMixin
 from Products.CMFPlone.utils import base_hasattr
-from Products.PageTemplates.Expressions import SecureModuleImporter
 from Products.PloneMeeting.browser.itemchangeorder import _compute_value_to_add
 from Products.PloneMeeting.browser.itemchangeorder import _is_integer
 from Products.PloneMeeting.browser.itemchangeorder import _to_integer
@@ -58,6 +56,7 @@ from Products.PloneMeeting.config import READER_USECASES
 from Products.PloneMeeting.interfaces import IMeetingWorkflowActions
 from Products.PloneMeeting.interfaces import IMeetingWorkflowConditions
 from Products.PloneMeeting.utils import _addManagedPermissions
+from Products.PloneMeeting.utils import _base_extra_expr_ctx
 from Products.PloneMeeting.utils import addDataChange
 from Products.PloneMeeting.utils import addRecurringItemsIfRelevant
 from Products.PloneMeeting.utils import display_as_html
@@ -891,14 +890,11 @@ class Meeting(OrderedBaseFolder, BrowserDefaultMixin):
         catalog = api.portal.get_tool('portal_catalog')
         tool = api.portal.get_tool('portal_plonemeeting')
         cfg = tool.getMeetingConfig(self)
-        # add GMT+x value
-        localizedValue0 = value + ' ' + _findLocalTimeZoneName(0)
-        localizedValue1 = value + ' ' + _findLocalTimeZoneName(1)
-        # make sure we look for existing meetings in both possible
-        # time zones because if we just changed timezone, we could create
-        # another meeting with same date of an existing that was created with previous timezone...
-        otherMeetings = catalog(portal_type=cfg.getMeetingTypeName(),
-                                getDate=(DateTime(localizedValue0), DateTime(localizedValue1), ))
+        # DateTime('2020-05-28 11:00') will result in DateTime('2020/05/28 11:00:00 GMT+0')
+        # DateTime('2020/05/28 11:00') will result in DateTime('2020/05/28 11:00:00 GMT+2')
+        # so make sure value uses "/" instead "-"
+        value = value.replace("-", "/")
+        otherMeetings = catalog(portal_type=cfg.getMeetingTypeName(), getDate=DateTime(value))
         if otherMeetings:
             for m in otherMeetings:
                 if m.getObject() != self:
@@ -1809,20 +1805,17 @@ class Meeting(OrderedBaseFolder, BrowserDefaultMixin):
 
     def _updatePowerObserversLocalRoles(self):
         '''Give local roles to the groups defined in MeetingConfig.powerObservers.'''
-        tool = api.portal.get_tool('portal_plonemeeting')
-        cfg = tool.getMeetingConfig(self)
-        itemState = self.queryState()
+        extra_expr_ctx = _base_extra_expr_ctx(self)
+        extra_expr_ctx.update({'meeting': self, })
+        cfg = extra_expr_ctx['cfg']
+        cfg_id = cfg.getId()
+        meetingState = self.queryState()
         for po_infos in cfg.getPowerObservers():
-            if itemState in po_infos['meeting_states'] and \
+            if meetingState in po_infos['meeting_states'] and \
                _evaluateExpression(self,
                                    expression=po_infos['meeting_access_on'],
-                                   extra_expr_ctx={
-                                       'meeting': self,
-                                       'pm_utils': SecureModuleImporter['Products.PloneMeeting.utils'],
-                                       'imio_history_utils': SecureModuleImporter['imio.history.utils'],
-                                       'tool': tool,
-                                       'cfg': cfg}):
-                powerObserversGroupId = "%s_%s" % (cfg.getId(), po_infos['row_id'])
+                                   extra_expr_ctx=extra_expr_ctx):
+                powerObserversGroupId = "%s_%s" % (cfg_id, po_infos['row_id'])
                 self.manage_addLocalRoles(powerObserversGroupId, (READER_USECASES['powerobservers'],))
 
     security.declareProtected(ModifyPortalContent, 'transformRichTextField')