+++ CHANGES.rst
@@ -7,7 +7,7 @@ Changelog
 
 - Use `Products.Archetypes.interfaces.IObjectInitializedEvent` instead `zope.lifecycleevent.interfaces.IObjectAddedEvent`
   to initialize freshly created item to make `plone.restapi` happy or item is not initialized and attributes
-  like `adviceIndex` are not added
+  like `adviceIndex` are not added.  With `plone.restapi`, validation is done before `ObjectInitializedEvent` is called
   [gbastien]
 
 4.1.25.1 (2020-06-02)