+++ CHANGES.rst
@@ -5,9 +5,9 @@ Changelog
 4.1.26 (unreleased)
 -------------------
 
-- Use `Products.Archetypes.interfaces.IObjectInitializedEvent` instead `zope.lifecycleevent.interfaces.IObjectAddedEvent`
+- Use `Products.Archetypes.interfaces.IObjectInitializedEvent` and `zope.lifecycleevent.interfaces.IObjectAddedEvent`
   to initialize freshly created item to make `plone.restapi` happy or item is not initialized and attributes
-  like `adviceIndex` are not added.  With `plone.restapi`, validation is done before `ObjectInitializedEvent` is called
+  like `adviceIndex` are not added.  With `plone.restapi`, validation is done after `ObjectInitializedEvent` but before `ObjectAddedEvent`
   [gbastien]
 - Added missing icon `wait_advices_from.png`
 - Do not fail in `vocabularies.PMCategoryVocabulary` when creating an annex using `plone.restapi`,